Rename permissions.js to unsafe.js
